+++ src/redirection/redirection.c
@@ -6,7 +6,7 @@
 /*   By: zmetreve <zmetreve@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 18:26:08 by zmetreve          #+#    #+#             */
-/*   Updated: 2025/07/01 02:34:39 by zmetreve         ###   ########.fr       */
+/*   Updated: 2025/07/04 22:20:34 by zmetreve         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,8 @@ bool	restore_io(t_io_fds *io)
 	}
 	return (ret);
 }
-/*
+
+
 bool	redirect_io(t_io_fds *io)
 {
 	int	ret;
@@ -60,37 +61,13 @@ bool	redirect_io(t_io_fds *io)
 	io->stdout_backup = dup(STDOUT_FILENO);
 	if (io->stdout_backup == -1)
 		ret = errmsg_cmd("dup", "stdout backup", strerror(errno), false);
-	if (io->fd_in != -1)
-		if (dup2(io->fd_in, STDIN_FILENO) == -1)
-			ret = errmsg_cmd("dup2", io->infile, strerror(errno), false);
-	if (io->fd_out != -1)
-		if (dup2(io->fd_out, STDOUT_FILENO) == -1)
-			ret = errmsg_cmd("dup2", io->outfile, strerror(errno), false);
-	return (ret);
-}*/
-bool	redirect_io(t_io_fds *io)
-{
-	int	ret;
-
-	ret = true;
-	if (!io)
-		return (ret);
-	io->stdin_backup = dup(STDIN_FILENO);
-	if (io->stdin_backup == -1)
-		ret = errmsg_cmd("dup", "stdin backup", strerror(errno), false);
-	io->stdout_backup = dup(STDOUT_FILENO);
-	if (io->stdout_backup == -1)
-		ret = errmsg_cmd("dup", "stdout backup", strerror(errno), false);
-
 	if (io->fd_in != -1)
 	{
-		printf("DEBUG: redirect_io -> duplicando fd_in %d a STDIN\n", io->fd_in);
 		if (dup2(io->fd_in, STDIN_FILENO) == -1)
 			ret = errmsg_cmd("dup2", io->infile, strerror(errno), false);
 	}
 	if (io->fd_out != -1)
 	{
-		printf("DEBUG: redirect_io -> duplicando fd_out %d a STDOUT\n", io->fd_out);
 		if (dup2(io->fd_out, STDOUT_FILENO) == -1)
 			ret = errmsg_cmd("dup2", io->outfile, strerror(errno), false);
 	}
